feat(k8s-core): finalize Kueue sentinel layout per audited plan
Move VERSIONS and L2/L3 scripts into package scripts/, split Trainer CRDs
into model-training, add shared crd-fetch lib, and expose kueue:test:api
and kueue:test:integration npm tasks for CI and local validation.
