+++ packages/k8s-core/scripts/test-manifests/clustertrainingruntime.yaml
@@ -1,0 +1,25 @@
+apiVersion: trainer.kubeflow.org/v1alpha1
+kind: ClusterTrainingRuntime
+metadata:
+  name: sentinel-runtime
+spec:
+  mlPolicy:
+    numNodes: 1
+    torch:
+      numProcPerNode: 1
+  template:
+    spec:
+      replicatedJobs:
+        - name: node
+          replicas: 1
+          template:
+            spec:
+              template:
+                spec:
+                  containers:
+                    - name: node
+                      image: registry.k8s.io/pause:3.9
+                      resources:
+                        requests:
+                          cpu: "50m"
+                          memory: 64Mi
