+++ packages/k8s-core/scripts/test-manifests/localqueue.yaml
@@ -1,0 +1,7 @@
+apiVersion: kueue.x-k8s.io/v1beta2
+kind: LocalQueue
+metadata:
+  name: sentinel-local-queue
+  namespace: kueue-sentinel
+spec:
+  clusterQueue: sentinel-cluster-queue
