+++ packages/k8s-core/scripts/test-manifests/workload-pause.yaml
@@ -1,0 +1,20 @@
+apiVersion: kueue.x-k8s.io/v1beta2
+kind: Workload
+metadata:
+  name: sentinel-workload
+  namespace: kueue-sentinel
+spec:
+  active: true
+  queueName: sentinel-local-queue
+  podSets:
+    - count: 1
+      name: main
+      template:
+        spec:
+          containers:
+            - name: pause
+              image: registry.k8s.io/pause:3.9
+              resources:
+                requests:
+                  cpu: "100m"
+                  memory: 128Mi
