+++ packages/k8s-core/src/__tests__/fixtures/kueue-crds/kueue.x-k8s.io_workloads.yaml
@@ -91,7 +91,7 @@ spec:
                 description: |-
                   podSets is a list of sets of homogeneous pods, each described by a Pod spec
                   and a count.
-                  There must be at least one element and at most 18.
+                  There must be at least one element and at most 10.
                   podSets cannot be changed.
                 items:
                   properties:
@@ -8796,7 +8796,7 @@ spec:
                   x-kubernetes-validations:
                   - message: minCount should be positive and less or equal to count
                     rule: 'has(self.minCount) ? self.minCount <= self.count : true'
-                maxItems: 18
+                maxItems: 10
                 minItems: 1
                 type: array
                 x-kubernetes-list-map-keys:
@@ -9016,7 +9016,7 @@ spec:
                       required:
                       - name
                       type: object
-                    maxItems: 18
+                    maxItems: 10
                     type: array
                     x-kubernetes-list-map-keys:
                     - name
@@ -9141,7 +9141,7 @@ spec:
                         required:
                         - name
                         type: object
-                      maxItems: 18
+                      maxItems: 10
                       type: array
                       x-kubernetes-list-type: atomic
                     requeueAfterSeconds:
@@ -9296,7 +9296,7 @@ spec:
                   - count
                   - name
                   type: object
-                maxItems: 18
+                maxItems: 10
                 type: array
                 x-kubernetes-list-map-keys:
                 - name
@@ -9353,7 +9353,7 @@ spec:
                   required:
                   - name
                   type: object
-                maxItems: 18
+                maxItems: 10
                 type: array
                 x-kubernetes-list-map-keys:
                 - name
@@ -9526,7 +9526,7 @@ spec:
                 description: |-
                   podSets is a list of sets of homogeneous pods, each described by a Pod spec
                   and a count.
-                  There must be at least one element and at most 18.
+                  There must be at least one element and at most 10.
                   podSets cannot be changed.
                 items:
                   properties:
@@ -18275,7 +18275,7 @@ spec:
                   x-kubernetes-validations:
                   - message: minCount should be positive and less or equal to count
                     rule: 'has(self.minCount) ? self.minCount <= self.count : true'
-                maxItems: 18
+                maxItems: 10
                 minItems: 1
                 type: array
                 x-kubernetes-list-map-keys:
@@ -18650,7 +18650,7 @@ spec:
                               number of levels in this TopologyAssignment
                             rule: self.slices.all(x, size(x.valuesPerLevel) == size(self.levels))
                       type: object
-                    maxItems: 18
+                    maxItems: 10
                     type: array
                     x-kubernetes-list-map-keys:
                     - name
@@ -18774,7 +18774,7 @@ spec:
                         required:
                         - name
                         type: object
-                      maxItems: 18
+                      maxItems: 10
                       type: array
                       x-kubernetes-list-type: atomic
                     requeueAfterSeconds:
@@ -18965,7 +18965,7 @@ spec:
                   - count
                   - name
                   type: object
-                maxItems: 18
+                maxItems: 10
                 type: array
                 x-kubernetes-list-map-keys:
                 - name
@@ -19022,7 +19022,7 @@ spec:
                   required:
                   - name
                   type: object
-                maxItems: 18
+                maxItems: 10
                 type: array
                 x-kubernetes-list-map-keys:
                 - name
