+++ packages/model-training/src/__tests__/fixtures/trainer-crds/trainer.kubeflow.org_clustertrainingruntimes.yaml
