+++ packages/model-training/src/__tests__/trainerContract.spec.ts
@@ -1,18 +1,24 @@
+import path from 'path';
+
 import {
-  fixturePath,
+  assertModelVersionServed,
   getVersionSchema,
   loadCrd,
   resolveSchemaPath,
   schemaSupportsType,
 } from '@odh-dashboard/k8s-core/__tests__/helpers/crdSchemaHelpers';
 
+const TRAINER_VERSION = 'v1alpha1';
+const TRAINER_API_GROUP = 'trainer.kubeflow.org';
+
 const TRAIN_JOB_MODEL = {
+  apiGroup: TRAINER_API_GROUP,
   apiVersion: 'v1alpha1',
-  apiGroup: 'trainer.kubeflow.org',
   kind: 'TrainJob',
 } as const;
 
-const TRAINER_VERSION = 'v1alpha1';
+const fixturePath = (...segments: string[]): string =>
+  path.join(__dirname, 'fixtures', 'trainer-crds', ...segments);
 
 type ContractField = {
   path: string;
@@ -31,15 +37,25 @@ const assertFieldsExist = (crdFile: string, fields: ContractField[]): void => {
 };
 
 describe('Trainer CRD contract tests', () => {
-  describe('TrainJobKind contract', () => {
-    it('TrainJobModel targets a served CRD version', () => {
-      const crd = loadCrd(fixturePath('trainer.kubeflow.org_trainjobs.yaml'));
-      const version = crd.spec.versions.find((entry) => entry.name === TRAIN_JOB_MODEL.apiVersion);
-      expect(version?.served).toBe(true);
-      expect(crd.spec.group).toBe(TRAIN_JOB_MODEL.apiGroup);
-      expect(crd.spec.names.kind).toBe(TRAIN_JOB_MODEL.kind);
+  describe('API models', () => {
+    it('TrainJob model targets a served CRD version', () => {
+      assertModelVersionServed(
+        fixturePath('trainer.kubeflow.org_trainjobs.yaml'),
+        TRAIN_JOB_MODEL.apiVersion,
+        TRAIN_JOB_MODEL.apiGroup,
+      );
     });
 
+    it('ClusterTrainingRuntime targets a served CRD version', () => {
+      assertModelVersionServed(
+        fixturePath('trainer.kubeflow.org_clustertrainingruntimes.yaml'),
+        TRAINER_VERSION,
+        TRAINER_API_GROUP,
+      );
+    });
+  });
+
+  describe('TrainJobKind contract', () => {
     it('validates dashboard-used TrainJob fields against CRD schema', () => {
       assertFieldsExist(fixturePath('trainer.kubeflow.org_trainjobs.yaml'), [
         { path: 'spec.runtimeRef', type: 'object' },
@@ -55,14 +71,22 @@ describe('Trainer CRD contract tests', () => {
         { path: 'status.jobsStatus', type: 'array' },
       ]);
     });
+
+    it('TrainJob CRD exposes metadata object for queue labels', () => {
+      const schema = getVersionSchema(
+        loadCrd(fixturePath('trainer.kubeflow.org_trainjobs.yaml')),
+        TRAINER_VERSION,
+      );
+      const metadataSchema = resolveSchemaPath(schema, 'metadata');
+      expect(metadataSchema).toBeDefined();
+      expect(schemaSupportsType(metadataSchema, 'object')).toBe(true);
+    });
   });
 
   describe('ClusterTrainingRuntimeKind contract', () => {
     it('validates dashboard-used ClusterTrainingRuntime fields against CRD schema', () => {
       assertFieldsExist(fixturePath('trainer.kubeflow.org_clustertrainingruntimes.yaml'), [
         { path: 'spec.mlPolicy', type: 'object' },
-        { path: 'spec.mlPolicy.numNodes', type: 'integer' },
-        { path: 'spec.mlPolicy.torch', type: 'object' },
         { path: 'spec.template', type: 'object' },
         { path: 'spec.template.spec', type: 'object' },
         { path: 'spec.template.spec.replicatedJobs', type: 'array' },
